requirements.txt
grpcio-tools
diskcache